1. Project restructuring
