+++ src/components/common/search.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import TextField from "./textField";
+import TextField from "../common/form/textField";
 import PropTypes from "prop-types";
 
 const Search = ({ onSearch, value }) => {
