+++ src/components/common/table/index.js
@@ -1,0 +1,6 @@
+import Table from './table';
+export default Table;
+
+import TableBody from "./tableBody";
+import TableHeader from "./tableHeader";
+export { TableBody, TableHeader }
