+++ src/components/page/userListPage/index.js
@@ -1,0 +1,3 @@
+import UserListPage from "./usersList";
+
+export default UserListPage;
