+++ src/components/ui/loginForm.jsx
@@ -1,0 +1,91 @@
+import React, { useEffect, useState } from "react";
+import { validator } from "../../utils/validator";
+import TextField from "../common/form/textField";
+
+const LoginForm = () => {
+    const [data, setData] = useState({
+        email: "",
+        password: ""
+    });
+    const [error, setError] = useState({});
+    const handleChange = ({ target }) => {
+        setData((prevState) => ({
+            ...prevState,
+            [target.name]: target.value
+        }));
+    };
+    const validatorConfig = {
+        email: {
+            isRequired: {
+                message: "Электронная почта обязательна для заполнения"
+            },
+            isEmail: {
+                message: "Email введен некорректно"
+            }
+        },
+        password: {
+            isRequired: {
+                message: "Пароль обязательна для заполнения"
+            },
+            isCapitalSymbol: {
+                message: "Пароль должен содержать хотя бы одну заглавную букву"
+            },
+            isContainDigit: {
+                message: "Пароль должен содержать хотя бы одно число"
+            },
+            min: {
+                message: "Пароль должен состоять минимум из 8 символов",
+                value: 8
+            }
+        }
+    };
+    useEffect(() => {
+        validate();
+    }, [data]);
+    const validate = () => {
+        const error = validator(data, validatorConfig);
+        setError(error);
+        return Object.keys(error).length === 0;
+    };
+    const isValid = Object.keys(error).length === 0;
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        const isValid = validate();
+        if (isValid) return;
+        console.log(data);
+    };
+    return (
+        <div className="container mt-5">
+            <div className="row">
+                <div className="col-md-6 offset-md3 shadow p-4">
+                    <h3 className="mb-4">Login</h3>
+                    <form onSubmit={handleSubmit}>
+                        <TextField
+                            label="Электронная почта"
+                            name="email"
+                            value={data.email}
+                            onChange={handleChange}
+                            error={error.email}
+                        />
+                        <TextField
+                            label="Пароль"
+                            type="password"
+                            name="password"
+                            value={data.password}
+                            onChange={handleChange}
+                            error={error.password}
+                        />
+                        <button
+                            type="submit"
+                            disabled={!isValid}
+                            className="btn btn-primary w-100 mx-auto"
+                        >Submit</button>
+                    </form>
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default LoginForm;
