+++ src/components/ui/usersTable.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
-import BookMark from "./bookmark";
-import QualitiesList from "./qualitiesList";
-import Table from "./table";
 import { Link } from "react-router-dom";
+import BookMark from "../common/bookmark";
+import Qualities from "./qualitiel/qualitiesList";
+import Table from "../common/table/table";
 
 const UsersTable = ({ users, onSort, selectedSort, onToggleBookMark, onDelete, ...rest }) => {
     const columns = {
@@ -17,7 +17,7 @@ const UsersTable = ({ users, onSort, selectedSort, onToggleBookMark, onDelete, .
         qualities: {
             name: "Качества",
             component: (user) => (
-                <QualitiesList qualities={user.qualities}/>
+                <Qualities qualities={user.qualities}/>
             )
         },
         professions: { path: "profession.name", name: "Профессия" },
