+++ src/layouts/users.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import UsersList from "../components/usersList";
 import { useParams } from "react-router-dom";
-import UserPage from "../components/userPage";
+import UsersListPage from "../components/page/userListPage";
+import UserPage from "../components/page/userPage";
 
 const Users = () => {
     const params = useParams();
@@ -10,7 +10,7 @@ const Users = () => {
         <>
             {userId
                 ? <UserPage userId={userId}/>
-                : <UsersList/>
+                : <UsersListPage/>
             }
         </>
     );
